Add "Refresh Controllers" button to main menu issue/#2125 and issue/#3648
* A "Refresh Controllers" button was added to main menu that partially
solves "recognize disconnect and reconnect" feature in issue/#2125. When
this button is pressed, the list of controllers are re-fetched to be
recognized by the system. This partially solves the problem of having
to close and re-open the game every time a controller is created.

* Also, when this button is pressed, the "Controller Settings" screen is
disposed from the assets. This is done to solve to force the system to re-create
the "Controller Settings" screen in case it had already loaded in the assets
before the "Refresh Controllers" was pressed. This makes the system show the controllers
in the "Controller Settings" screen after re-fetching the controllers.
That is why it is related to issue/#3648.

Note that: the re-fetching of controllers was implemented and tested only for windows 10
and Xbox 360 controller. Implementing and testing needs to be done for other operating systems
and controllers.